+++ CHANGES.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added `Loretta.CodeAnalysis.Lua.Experimental.SyntaxExtensions.ConstantFold`.
+
+### Changed
+- DEPRECATED: `Loretta.CodeAnalysis.Lua.Experimental.SyntaxExtensions.FoldConstants` in favor of `ConstantFold`.
 
 ## v0.2.7-beta.6
 ### Added